+++ src/main/java/com/chrisrm/ideaddlc/utils/MTUiUtils.java
@@ -199,7 +199,7 @@ public enum MTUiUtils {
   public static String getVersion() {
      return getPlugin()
              .map(IdeaPluginDescriptor::getVersion)
-             .orElse("2.4.1");
+             .orElse("2.5.0");
   }
 
   private static Optional<String> getPluginId() {