+++ src/main/java/io/acari/DDLC/themes/light/MonikaTheme.kt
@@ -93,6 +93,10 @@ class MonikaTheme : DokiDokiTheme("monika", "Monika", false, "Monika") {
 
     override fun getBorderColor(): Color = Color.getHSBColor(62f, 91f, 149f)
 
+    override fun getStartColor(): String = "66B757"
+
+    override fun getStopColor(): String = "64FF1D"
+
 
     companion object {
         val BACKGROUND = "fffcfc"