+++ docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 ----
 
+# 2.5.0
+
+- Class icons are now themed to the current club member.
+- Changed function icons to be λ
+-   
+
+
 # 2.4.1
 
 - Updated plugin display icon.